refactor(follow): CursorBasedPaginatedResult 위치 이동
#13
